Move hasSolicitorAgreedToReceiveEmails to CommonFunction class
