+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/helper/CommonConditions.java
@@ -4,13 +4,11 @@ import org.apache.commons.lang3.StringUtils;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.RespondToOrderData;
 
 import java.util.Map;
-import java.util.Objects;
 
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.AMENDED_CONSENT_ORDER;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.APPLICANT_REPRESENTED;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.RESP_SOLICITOR_NAME;
-import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.SOLICITOR_AGREE_TO_RECEIVE_EMAILS;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.service.CommonFunction.nullToEmpty;
 
 public class CommonConditions {
@@ -28,10 +26,5 @@ public class CommonConditions {
 
     public static boolean isAmendedConsentOrderType(RespondToOrderData respondToOrderData) {
         return AMENDED_CONSENT_ORDER.equalsIgnoreCase(respondToOrderData.getRespondToOrder().getDocumentType());
-    }
-
-    public static boolean hasSolicitorAgreedToReceiveEmails(Map<String, Object> mapOfCaseData) {
-        return YES_VALUE.equalsIgnoreCase(Objects.toString(mapOfCaseData
-            .get(SOLICITOR_AGREE_TO_RECEIVE_EMAILS)));
     }
 }
